👌 IMP: Switch to just use std::sync::atomic

+++ src/options.rs
@@ -1,5 +1,5 @@
-use atomics::{AtomicU64, AtomicUsize, Ordering};
 use std::cmp::max;
+use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 
 static NUM_THREADS: AtomicUsize = AtomicUsize::new(1);
 static HASH_SIZE_MB: AtomicUsize = AtomicUsize::new(16);
@@ -22,10 +22,10 @@ pub fn get_hash_size_mb() -> usize {
     max(1, HASH_SIZE_MB.load(Ordering::Relaxed))
 }
 
-pub fn set_exploration_constant(c: usize) {
+pub fn set_exploration_constant(c: u64) {
     EXPLORATION_CONSTANT.store(c, Ordering::Relaxed);
 }
 
-pub fn get_exploration_constant() -> usize {
+pub fn get_exploration_constant() -> u64 {
     EXPLORATION_CONSTANT.load(Ordering::Relaxed)
 }
